style&feat: 기타 디자인 변경, 메인페이지 여행계획에 추가시 오류 모달
