+++ client/src/components/MainPage/MainMap.js
@@ -34,7 +34,7 @@ const MainMap = () => {
   return mapCenter?.length !== 0 ? (
     <div className="w-full h-60 sm:h-[600px] flex justify-center">
       <Map // 지도를 표시할 Container
-        className="border-2 border-[#59AEEC] rounded-xl "
+        className="border-2 border-[#59AEEC] rounded-xl"
         center={{ lat: mapCenter[0]?.lat, lng: mapCenter[0]?.lng }}
         style={{
           // 지도의 크기
@@ -51,31 +51,31 @@ const MainMap = () => {
         {postList?.length !== 0 ? (
           postList?.map((data, index) => {
             const detailData = {
-              title: data.title,
-              star: data.star,
-              tags: data.tags,
+              title: data?.title,
+              star: data?.star,
+              tags: data?.tags,
             };
             return (
               // 커스텀 오버레이를 표시할 Container
               <CustomOverlayMap
                 key={index}
                 // 커스텀 오버레이가 표시될 위치입니다
-                position={{ lat: data.latitude, lng: data.longitude }}
+                position={{ lat: data?.latitude, lng: data?.longitude }}
                 // 커스텀 오버레이가에 대한 확장 옵션 x,y 좌표 이동.
-                xAnchor={mapImgClickId === data.boardId ? 0.05 : 0.1}
-                yAnchor={mapImgClickId === data.boardId ? 1.0 : 0.91}
+                xAnchor={mapImgClickId === data?.boardId ? 0.05 : 0.1}
+                yAnchor={mapImgClickId === data?.boardId ? 1.0 : 0.91}
                 // zIndex : z-index
-                zIndex={handelZIndex(data.boardId)}
+                zIndex={handelZIndex(data?.boardId)}
               >
                 <div
-                  onMouseEnter={() => setMapImgHoverId(data.boardId)}
+                  onMouseEnter={() => setMapImgHoverId(data?.boardId)}
                   onMouseLeave={() => setMapImgHoverId(null)}
                 >
                   <MapItem
-                    thumbnail={data.thumbnail}
-                    boardId={data.boardId}
+                    thumbnail={data?.thumbnail}
+                    boardId={data?.boardId}
                     detailData={detailData}
-                    coordinate={{ lat: data.latitude, lng: data.longitude }}
+                    coordinate={{ lat: data?.latitude, lng: data?.longitude }}
                   />
                 </div>
               </CustomOverlayMap>
