+++ client/src/components/homePage/HomeMap.js
@@ -71,7 +71,12 @@ export default function HomeMap() {
           </div>
         )}
       </div>
-      <Map level={14} center={isCenter} style={style} className="lg:m-0 m-auto">
+      <Map
+        level={13}
+        center={isCenter}
+        style={style}
+        className="lg:m-0 m-auto border-2 border-[#59AEEC] rounded-xl"
+      >
         <Polyline
           path={[
             [
