+++ client/src/components/mytravel/MyTravelMap.js
@@ -1,5 +1,5 @@
 import axios from "axios";
-import { useEffect, useState } from "react";
+import { useEffect, useRef, useState } from "react";
 import { TiPencil } from "react-icons/ti";
 import { Map, CustomOverlayMap, Polyline } from "react-kakao-maps-sdk";
 import { useRecoilState, useRecoilValue } from "recoil";
@@ -13,6 +13,7 @@ import {
 // import TravelMapItem from "./MapItem/TravelMapItem";
 
 const MyTravelMap = ({ data }) => {
+  const focusRef = useRef();
   const [TOKEN] = useRecoilState(accessToken);
   const [, setMyTravelName] = useRecoilState(myTravelNameSelect);
 
@@ -77,7 +78,8 @@ const MyTravelMap = ({ data }) => {
           <>
             <input
               type="text"
-              className="h-7 pl-2 outline-none border-b-2 border-[rgb(83,199,240)] text-[rgb(83,199,240)] text-lg font-semibold z-20"
+              ref={focusRef}
+              className="min-w-xs h-8  outline-none border-b-2 border-[rgb(83,199,240)] text-[rgb(83,199,240)] text-lg font-semibold z-20"
               defaultValue={myTravelName}
               onChange={(e) => setInputValue(e.target.value)}
             ></input>
@@ -89,7 +91,7 @@ const MyTravelMap = ({ data }) => {
               }`}
               onClick={changeName}
             >
-              <TiPencil className="" size={"25"} />
+              <TiPencil className="" size={"29"} />
             </button>
             <button
               className="fixed inset-0 z-10 cursor-default"
@@ -99,14 +101,14 @@ const MyTravelMap = ({ data }) => {
           </>
         ) : (
           <>
-            <h2 className="text-lg font-semibold text-[rgb(83,199,240)]">
+            <h2 className="max-w-xs truncate text-lg font-semibold text-[rgb(83,199,240)]">
               {preViewText ? preViewText : myTravelName}
             </h2>
             <button
               className={`ml-2 text-[rgb(83,199,240)] border-b-2 border-[rgb(83,199,240)] `}
               onClick={() => setInputText(!inputText)}
             >
-              <TiPencil size={"25"} />
+              <TiPencil size={"29"} />
             </button>
           </>
         )}
